workingdir/requirements-full.txt
absl-py==1.2.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asn1crypto==0.24.0
astor==0.8.1
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
carla @ file:///home/carla/carla/PythonAPI/carla/dist/carla-0.9.13-cp37-cp37m-linux_x86_64.whl
certifi==2018.1.18
cffi==1.14.0
chardet==3.0.4
charset-normalizer==2.1.1
contextlib2==21.6.0
cryptography==2.1.4
cycler==0.11.0
Cython==0.29.15
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distro==1.7.0
docopt==0.6.2
entrypoints==0.4
fastjsonschema==2.16.2
fonttools==4.37.4
future==0.18.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.49.1
h5py==2.10.0
httplib2==0.9.2
idna==2.6
importlib-metadata==5.0.0
importlib-resources==5.9.0
ipykernel==6.16.0
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.4.4
llvmlite==0.33.0
lucid==0.3.1
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.4
mock==4.0.3
nbclient==0.6.8
nbconvert==7.1.0
nbformat==5.6.1
nest-asyncio==1.5.6
networkx==2.6.3
notebook==6.4.12
numba==0.50.1
numpy==1.19.0
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==21.3
pandas==1.0.5
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pipreqs==0.4.11
pkgutil_resolve_name==1.3.10
platformdirs==3.1.1
pocl-binary-distribution==3.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.19.0
psutil==5.9.2
ptyprocess==0.7.0
pycapnp==0.6.4
pycparser==2.21
pycrypto==2.6.1
pygame==2.1.2
Pygments==2.13.0
PyGObject==3.26.1
pymap3d==2.4.1
pyopencl==2019.1.2
PyOpenGL==3.1.6
pyparsing==3.0.9
pyrsistent==0.18.1
python-apt==1.6.5+ubuntu0.7
python-dateutil==2.8.2
pytools==2022.1.12
pytz==2022.7.1
pyxdg==0.25
PyYAML==5.3.1
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
requests==2.28.1
requests-unixsocket==0.1.5
scikit-learn==1.0.2
scipy==1.7.3
SecretStorage==2.3.1
Send2Trash==1.8.0
Shapely==1.6.4.post2
six==1.16.0
soupsieve==2.3.2.post1
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==2.0.1
terminado==0.16.0
threadpoolctl==3.1.0
tinycss2==1.1.1
torch==1.12.1+cu116
torchaudio==0.12.1+cu116
torchvision==0.13.1+cu116
tornado==6.2
tqdm==4.65.0
traitlets==5.4.0
typing_extensions==4.5.0
unattended-upgrades==0.1
urllib3==1.22
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.15.0
yarg==0.1.9
zipp==3.8.1
zmq==0.0.0
